+++ app/keepers/keepers.go
@@ -54,8 +54,6 @@ import (
 	epochskeeper "github.com/osmosis-labs/osmosis/v10/x/epochs/keeper"
 	epochstypes "github.com/osmosis-labs/osmosis/v10/x/epochs/types"
 	gammkeeper "github.com/osmosis-labs/osmosis/v10/x/gamm/keeper"
-	"github.com/osmosis-labs/osmosis/v10/x/gamm/twap"
-	twaptypes "github.com/osmosis-labs/osmosis/v10/x/gamm/twap/types"
 	gammtypes "github.com/osmosis-labs/osmosis/v10/x/gamm/types"
 	incentiveskeeper "github.com/osmosis-labs/osmosis/v10/x/incentives/keeper"
 	incentivestypes "github.com/osmosis-labs/osmosis/v10/x/incentives/types"
@@ -71,6 +69,8 @@ import (
 	superfluidtypes "github.com/osmosis-labs/osmosis/v10/x/superfluid/types"
 	tokenfactorykeeper "github.com/osmosis-labs/osmosis/v10/x/tokenfactory/keeper"
 	tokenfactorytypes "github.com/osmosis-labs/osmosis/v10/x/tokenfactory/types"
+	"github.com/osmosis-labs/osmosis/v10/x/twap"
+	twaptypes "github.com/osmosis-labs/osmosis/v10/x/twap/types"
 	"github.com/osmosis-labs/osmosis/v10/x/txfees"
 	txfeeskeeper "github.com/osmosis-labs/osmosis/v10/x/txfees/keeper"
 	txfeestypes "github.com/osmosis-labs/osmosis/v10/x/txfees/types"
@@ -330,13 +330,10 @@ func (appKeepers *AppKeepers) InitNormalKeepers(
 		appKeepers.BankKeeper,
 		appKeepers.IncentivesKeeper,
 		appKeepers.DistrKeeper,
-		distrtypes.ModuleName,
-		authtypes.FeeCollectorName,
 	)
 	appKeepers.PoolIncentivesKeeper = &poolIncentivesKeeper
 
 	tokenFactoryKeeper := tokenfactorykeeper.NewKeeper(
-		appCodec,
 		appKeepers.keys[tokenfactorytypes.StoreKey],
 		appKeepers.GetSubspace(tokenfactorytypes.ModuleName),
 		appKeepers.AccountKeeper,
@@ -349,7 +346,7 @@ func (appKeepers *AppKeepers) InitNormalKeepers(
 	// if we want to allow any custom callbacks
 	supportedFeatures := "iterator,staking,stargate,osmosis"
 
-	wasmOpts = append(owasm.RegisterCustomPlugins(appKeepers.GAMMKeeper, appKeepers.BankKeeper, appKeepers.TokenFactoryKeeper), wasmOpts...)
+	wasmOpts = append(owasm.RegisterCustomPlugins(appKeepers.GAMMKeeper, appKeepers.BankKeeper, appKeepers.TwapKeeper, appKeepers.TokenFactoryKeeper), wasmOpts...)
 
 	wasmKeeper := wasm.NewKeeper(
 		appCodec,