+++ app/modules.go
@@ -51,8 +51,6 @@ import (
 	"github.com/osmosis-labs/osmosis/v10/x/epochs"
 	epochstypes "github.com/osmosis-labs/osmosis/v10/x/epochs/types"
 	"github.com/osmosis-labs/osmosis/v10/x/gamm"
-	"github.com/osmosis-labs/osmosis/v10/x/gamm/twap"
-	twaptypes "github.com/osmosis-labs/osmosis/v10/x/gamm/twap/types"
 	gammtypes "github.com/osmosis-labs/osmosis/v10/x/gamm/types"
 	"github.com/osmosis-labs/osmosis/v10/x/incentives"
 	incentivestypes "github.com/osmosis-labs/osmosis/v10/x/incentives/types"
@@ -66,6 +64,8 @@ import (
 	superfluidtypes "github.com/osmosis-labs/osmosis/v10/x/superfluid/types"
 	"github.com/osmosis-labs/osmosis/v10/x/tokenfactory"
 	tokenfactorytypes "github.com/osmosis-labs/osmosis/v10/x/tokenfactory/types"
+	"github.com/osmosis-labs/osmosis/v10/x/twap"
+	twaptypes "github.com/osmosis-labs/osmosis/v10/x/twap/types"
 	"github.com/osmosis-labs/osmosis/v10/x/txfees"
 	txfeestypes "github.com/osmosis-labs/osmosis/v10/x/txfees/types"
 )