+++ x/mint/keeper/keeper_test.go
@@ -14,7 +14,7 @@ import (
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 
 	"github.com/osmosis-labs/osmosis/v10/app/apptesting"
-	"github.com/osmosis-labs/osmosis/v10/osmoutils"
+	"github.com/osmosis-labs/osmosis/v10/app/apptesting/osmoassert"
 	"github.com/osmosis-labs/osmosis/v10/x/mint/keeper"
 	"github.com/osmosis-labs/osmosis/v10/x/mint/types"
 	poolincentivestypes "github.com/osmosis-labs/osmosis/v10/x/pool-incentives/types"
@@ -431,7 +431,7 @@ func (suite *KeeperTestSuite) TestDistributeToModule() {
 	for name, tc := range tests {
 		suite.Run(name, func() {
 			suite.Setup()
-			osmoutils.ConditionalPanic(suite.T(), tc.expectPanic, func() {
+			osmoassert.ConditionalPanic(suite.T(), tc.expectPanic, func() {
 				mintKeeper := suite.App.MintKeeper
 				bankKeeper := suite.App.BankKeeper
 				accountKeeper := suite.App.AccountKeeper
@@ -667,7 +667,7 @@ func (suite *KeeperTestSuite) TestDistributeDeveloperRewards() {
 		suite.Run(name, func() {
 			suite.Setup()
 
-			osmoutils.ConditionalPanic(suite.T(), tc.expectPanic, func() {
+			osmoassert.ConditionalPanic(suite.T(), tc.expectPanic, func() {
 				mintKeeper := suite.App.MintKeeper
 				bankKeeper := suite.App.BankKeeper
 				accountKeeper := suite.App.AccountKeeper