+++ x/pool-incentives/types/key.go
@@ -20,10 +20,12 @@ var (
 	DistrInfoKey         = []byte("distr_info")
 )
 
+// GetPoolGaugeIdStoreKey returns a StoreKey with pool ID and its duration as inputs
 func GetPoolGaugeIdStoreKey(poolId uint64, duration time.Duration) []byte {
 	return []byte(fmt.Sprintf("pool-incentives/%d/%s", poolId, duration.String()))
 }
 
+// GetPoolIdFromGaugeIdStoreKey returns a StoreKey from the given gaugeID and duration.
 func GetPoolIdFromGaugeIdStoreKey(gaugeId uint64, duration time.Duration) []byte {
 	return []byte(fmt.Sprintf("pool-incentives-pool-id/%d/%s", gaugeId, duration.String()))
 }